requirementes.txt
Django==1.11.3
django-vote==2.1.4
pytz==2017.2
